+++ .editorconfig
@@ -8,7 +8,7 @@ trim_trailing_whitespace = true
 indent_style = space
 indent_size = 4
 
-[*.{md,yml,yaml,json}]
+[*.{md,yml,json,json5,toml}]
 indent_size = 2
 
 [*.luau]
