+++ .github/workflows/release.yml
@@ -19,13 +19,13 @@ jobs:
       fail-fast: false
       matrix:
         include:
-          - { target: x86_64-unknown-linux-gnu,  os: ubuntu-latest  }
+          - { target: x86_64-unknown-linux-gnu,  os: ubuntu-latest }
           - { target: aarch64-unknown-linux-gnu, os: ubuntu-24.04-arm }
-          - { target: x86_64-apple-darwin,       os: macos-13       }
-          - { target: aarch64-apple-darwin,      os: macos-latest   }
+          - { target: x86_64-apple-darwin,       os: macos-13 }
+          - { target: aarch64-apple-darwin,      os: macos-latest }
           - { target: x86_64-pc-windows-msvc,    os: windows-latest }
     steps:
-      - uses: actions/checkout@v4
+      - uses: actions/checkout@v5
       - run: rustup target add ${{ matrix.target }}
       - uses: Swatinem/rust-cache@v2
         with:
@@ -33,13 +33,10 @@ jobs:
 
       - run: cargo build --release --target ${{ matrix.target }}
 
-      # Zip on every platform, holding the binary alone at the archive root.
-      #
-      # Aftman reads zip only, and when no entry matches the tool name it installs the
-      # first one ending with the platform's executable suffix. That suffix is empty on
-      # Unix, so a README beside the binary would be installed as the tool. Rokit is
-      # stricter but scores a lone executable highest either way. The archive name
-      # carries the target triple, which is what both use to pick an asset.
+      # Zip everywhere, holding the binary alone at the archive root. Aftman reads zip
+      # only, and when no entry matches the tool name it installs the first one ending
+      # with the platform's executable suffix, which is empty on Unix. A README beside
+      # the binary would be installed as the tool.
       - name: Package
         shell: bash
         run: |
@@ -59,9 +56,7 @@ jobs:
       - uses: actions/upload-artifact@v4
         with:
           name: ${{ matrix.target }}
-          path: |
-            dist/*.zip
-            dist/*.sha256
+          path: dist/*
           if-no-files-found: error
 
   publish:
@@ -74,8 +69,6 @@ jobs:
           path: artifacts
           merge-multiple: true
 
-      # One combined checksum file, so a user can verify every asset with a single
-      # `sha256sum -c`.
       - name: Collect checksums
         run: |
           cd artifacts
@@ -84,10 +77,9 @@ jobs:
             rm "$file"
           done
           sort -k2 -o SHA256SUMS SHA256SUMS
-          cat SHA256SUMS
 
-      # Published as a draft: a release whose assets are still uploading is a release
-      # someone can download half of. Promote it once the assets are attached.
+      # Draft, because a release whose assets are still uploading is one someone can
+      # download half of.
       - uses: softprops/action-gh-release@v2
         with:
           draft: true
