+++ docs/SUMMARY.md
@@ -1,20 +1,8 @@
 # Table of contents
 
 * [ProCMP](README.md)
-
-## Start
-
 * [Install](install.md)
-* [Quickstart](quickstart.md)
-
-## Build
-
-* [Vars and defines](defines.md)
-* [darklua](darklua.md)
-* [Inputs](inputs.md)
-
-## Reference
-
 * [Manifest](manifest.md)
+* [darklua](darklua.md)
 * [CLI](cli.md)
 * [Diagnostics](diagnostics.md)
