+++ docs/darklua.md
@@ -4,32 +4,33 @@ description: darklua's configuration, passed through untouched.
 
 # darklua
 
-A profile's `darklua` block is darklua's own configuration format. ProCMP does not
-model it, rename it, restrict it or reorder it. The block is deserialised by darklua
-itself, so the accepted set is exactly what the linked version supports.
+A profile's `darklua` block is darklua's own configuration format. ProCMP does not model
+it, rename it, restrict it or reorder it. The block is deserialised by darklua itself,
+so the accepted set is exactly what the linked version supports, and a key it does not
+know is an error carrying darklua's own message and the JSON that was emitted.
 
-{% code title="pcmp.luau" %}
-```lua
-darklua = {
-	generator      = { name = "dense", column_span = 120 },
-	apply_to_files = { "src/**" },
-	skip_files     = { "**/*.test.luau" },
-	lua_extension  = "luau",
+{% code title="pcmp.json5" %}
+```json5
+darklua: {
+  generator: { name: "dense", column_span: 120 },
+  apply_to_files: ["src/**"],
+  skip_files: ["**/*.test.luau"],
+  lua_extension: "luau",
 
-	bundle = {
-		require_mode = { name = "luau", aliases = { pkg = "./Packages" } },
-		excludes     = { "@lune/**" },
-	},
+  bundle: {
+    require_mode: { name: "luau", aliases: { pkg: "./Packages" } },
+    excludes: ["@lune/**"],
+  },
 
-	rules = {
-		"compute_expression",
-		"remove_unused_if_branch",
-		"remove_types",
-		"remove_comments",
-		"rename_variables",
-		{ rule = "convert_require", current = "path", target = "roblox" },
-	},
-},
+  rules: [
+    "compute_expression",
+    "remove_unused_if_branch",
+    "remove_types",
+    "remove_comments",
+    "rename_variables",
+    { rule: "convert_require", current: "path", target: "roblox" },
+  ],
+}
 ```
 {% endcode %}
 
@@ -39,49 +40,33 @@ darklua = {
 
 ## Rules
 
-Three ways to say it, and they mean three different things:
+Three ways to say it, meaning three different things:
 
 <table><thead><tr><th width="200"></th><th></th></tr></thead><tbody>
 <tr><td><code>rules</code> omitted</td><td>darklua applies its own default rules</td></tr>
-<tr><td><code>rules = {}</code></td><td>No rules. Bundle and generate, transform nothing</td></tr>
-<tr><td><code>rules = { a, b, c }</code></td><td>Exactly these, in this order</td></tr>
+<tr><td><code>rules: []</code></td><td>No rules. Bundle and generate, transform nothing</td></tr>
+<tr><td><code>rules: [a, b, c]</code></td><td>Exactly these, in this order</td></tr>
 </tbody></table>
 
-```
-$ pcmp plan
-  release  dist/app.luau  darklua defaults
-```
-
 A rule in object form takes `apply_to_files` and `skip_files`, so one rule can be scoped
-without splitting the profile:
-
-```lua
-rules = {
-	"compute_expression",
-	{ rule = "remove_comments", skip_files = "**/vendor/**" },
-	{ rule = "rename_variables", apply_to_files = "src/**/*.luau" },
-},
-```
+without splitting the profile.
 
 ## Injection comes first
 
-Each [`define`](defines.md) becomes an `inject_global_value` rule, placed ahead of
-whatever you wrote. Nothing downstream can fold a value that has not been substituted
-yet, so this is the only position that works.
+Each `define` becomes an `inject_global_value` rule ahead of whatever you wrote. Nothing
+downstream can fold a value that has not been substituted yet.
 
 ```
-$ pcmp explain a
+$ pcmp explain release
 
 darklua configuration
   {
-    "apply_to_files": [ "src/**" ],
     "bundle": { "require_mode": { "name": "luau" } },
     "generator": "dense",
-    "lua_extension": "luau",
     "rules": [
       { "rule": "inject_global_value", "identifier": "DEBUG", "value": false },
       { "rule": "inject_global_value", "identifier": "PCMP_NAME", "value": "app" },
-      { "rule": "inject_global_value", "identifier": "PCMP_PROFILE", "value": "a" },
+      { "rule": "inject_global_value", "identifier": "PCMP_PROFILE", "value": "release" },
       "compute_expression",
       "remove_unused_if_branch"
     ]
@@ -103,14 +88,13 @@ error  branch-before-fold: task `release`: `remove_unused_if_branch` runs before
        help:   a branch is only removable once its condition has folded to a constant
 ```
 
-darklua's own default rule list is a valid ordering, so a lint stricter than the tool it
-lints for would reject working manifests. A rule listed twice is not a finding either,
-because running a pass again after an earlier rule has exposed new foldable code is a
-technique.
+darklua's own default list is a valid ordering, so a lint stricter than the tool it
+lints for would reject working manifests. A rule listed twice is not a finding either:
+running a pass again after an earlier rule exposed new foldable code is a technique.
 
 ## Headers
 
 `remove_comments` strips Luau directives along with everything else, and the `dense` and
-`readable` generators discard comments regardless. Headers are therefore written by
-ProCMP after darklua finishes, so nothing downstream can remove them. That is what the
-`header` field on a profile is for. See [Manifest](manifest.md).
+`readable` generators discard comments regardless. Headers are written by ProCMP after
+darklua finishes, so nothing downstream can remove them. See
+[`header`](manifest.md#entry-and-output).
