+++ docs/diagnostics.md
@@ -30,26 +30,17 @@ bad-template          an output or header token is unknown, empty, or unclosed
 bad-path              a path in entry, output or sources is empty, or escapes the root
 unknown-matrix-base   matrix base names a profile that does not exist
 empty-axis            a matrix axis has no values
-bad-define            a define key is not a Luau identifier, or its value is not finite
-bad-var               a var name cannot become a token and a constant, or two collide
 duplicate-axis-value  a matrix axis lists the same value twice
+bad-define            a define key is not a Luau identifier, or its value is not finite
+bad-var               a var name is not a name, or two collide as one constant
 bad-rules             darklua.rules is not a list darklua could read
 darklua-loaders       loaders declared in both `loaders` and `darklua.loaders`
 no-tasks              no profiles and no matrix
 ```
 
-`bad-var` covers both halves too. A name has to be writable as a Luau identifier, and
-two names that differ only in case would both become one `PCMP_<NAME>` constant, so the
-second would quietly replace the first.
-
-`bad-define` covers both halves of "this define cannot become a literal".
-`inject_global_value` substitutes by name, so a key such as `my-flag` or `end` matches
-nothing. Infinity and NaN have no literal form. A var and a matrix axis each contribute
-a `PCMP_<NAME>` constant, so their names are held to the same rule.
-
-`darklua-loaders` exists because only one of the two can win, and picking silently would
-mean the manifest format decided which loader pattern applied. See
-[Manifest](manifest.md#loaders).
+`inject_global_value` substitutes by name, so a define key such as `my-flag` or `end`
+matches nothing. A var becomes `PCMP_<NAME>` uppercased, so `channel` and `Channel`
+would both become one constant and the second would quietly replace the first.
 
 ## Rule order
 
@@ -59,9 +50,6 @@ Errors. Both are orderings darklua's own documentation calls out.
 fold-before-inject    compute_expression scheduled before inject_global_value
 branch-before-fold    remove_unused_if_branch scheduled before compute_expression
 ```
-
-The order is yours. These report rather than rewrite, because a silently reordered
-pipeline would make `pcmp explain` a lie. See [darklua](darklua.md).
 
 ## Hygiene
 
@@ -74,16 +62,6 @@ identical-profiles    two profiles declared identically
 
 ## Errors without a code
 
-Reported on stderr: two tasks claiming one output path, a missing entry point, a
-manifest that does not parse, an `ignore` entry that is not a valid glob, a configuration
-darklua rejected, which carries the JSON that was emitted, and a task filter that
-matched nothing.
-
-## Exit codes
-
-<table><thead><tr><th width="90">Code</th><th></th></tr></thead><tbody>
-<tr><td><code>0</code></td><td>Success</td></tr>
-<tr><td><code>1</code></td><td>A build task failed, or output was not reproducible</td></tr>
-<tr><td><code>2</code></td><td>The manifest could not be loaded or resolved</td></tr>
-<tr><td><code>5</code></td><td><code>check</code> found an error, or a warning under <code>--strict</code></td></tr>
-</tbody></table>
+Reported on stderr: two tasks claiming one output, a missing entry point, a manifest
+that does not parse, an `ignore` entry that is not a valid glob, a configuration darklua
+rejected, and a selection that matched nothing.
