+++ rustfmt.toml
@@ -1,4 +1,4 @@
-edition            = "2024"
-max_width          = 100
-newline_style      = "Unix"
+edition                  = "2024"
+max_width                = 100
+newline_style            = "Unix"
 use_field_init_shorthand = true
